back-end/betarequirements.txt
Django
Pillow
psycopg2
validate-email
python-dotenv
django-crispy-forms
social-auth-app-django
social-auth-core
django-sslserver
six
djangorestframework
qrcode
django-cors-headers
stripe
django_rest_passwordreset
djangorestframework
django-cors-headers
django-rest-passwordreset